+++ generators/state/templates/commonjs/state.js
@@ -1,6 +1,6 @@
 /*
  * <%- name %> state
- * <%- underline %>
+ * <%- '='.repeat(6 + name.length) %>
  *
 <% if (description) { -%>
  * <%- description %>
@@ -8,39 +8,9 @@
  */
 
 'use strict';
+<% methods.forEach(method => { -%>
 
-<% if (withInit) { -%>
-exports.init = function (/*game*/) {
+exports.<%- method %> = function (<%- method !== 'init' ? '/*game*/' : '' %>) {
   // TODO: Stub
 };
-
-<% } -%>
-<% if (withPreload) { -%>
-exports.preload = function (/*game*/) {
-  // TODO: Stub
-};
-
-<% } -%>
-<% if (withCreate) { -%>
-exports.create = function (/*game*/) {
-  // TODO: Stub
-};
-
-<% } -%>
-<% if (withUpdate) { -%>
-exports.update = function (/*game*/) {
-  // TODO: Stub
-};
-
-<% } -%>
-<% if (withRender) { -%>
-exports.render = function (/*game*/) {
-  // TODO: Stub
-};
-
-<% } -%>
-<% if (withShutdown) { -%>
-exports.shutdown = function (/*game*/) {
-  // TODO: Stub
-};
-<% } -%>
+<% }) -%>