+++ test/app.js
@@ -4,28 +4,58 @@
 
 'use strict';
 
+const chalk = require('chalk');
 const assert = require('yeoman-assert');
 const helpers = require('yeoman-test');
 
-const files = {
-  // Initial sample assets
-  assets: [
-    'static/assets/phaser.png',
-    'static/assets/progress-bar.png',
-    'static/assets/splash-screen.png',
-    'static/index.html'
-  ],
-  // Project configuration by base template
-  config: {
-    commonjs: [
+// User inputs.
+const title = 'My Test Game';
+const description = 'My awesome test game.';
+const customBuild = 'phaser';
+
+function runGenerator() {
+  return helpers
+    .run(require.resolve('../generators/app'));
+}
+
+describe(chalk.bold.cyan('generator-phaser-plus:app'), function () {
+  this.timeout(0);
+
+  describe('creates CommonJS-based projects', () => {
+    it('using prompts', () =>
+      runGenerator()
+        .withPrompts({title, description, customBuild, baseTemplate: 'commonjs'})
+        .then(checkAssets)
+        .then(checkReadme)
+        .then(checkGulpTasks)
+        .then(checkCommonJsConfig)
+        .then(checkCommonJsModules));
+  });
+
+  describe('creates ECMAScript-based projects', () => {
+    it('using prompts', () =>
+      runGenerator()
+        .withPrompts({title, description, customBuild, baseTemplate: 'esnext'})
+        .then(checkAssets)
+        .then(checkReadme)
+        .then(checkGulpTasks)
+        .then(checkECMAScriptConfig)
+        .then(checkECMAScriptModules));
+  });
+
+  function checkCommonJsConfig() {
+    assert.file([
       '.editorconfig',
       '.eslintrc.yml',
       '.gitattributes',
       '.gitignore',
       '.yo-rc.json',
       'package.json'
-    ],
-    esnext: [
+    ]);
+  }
+
+  function checkECMAScriptConfig() {
+    assert.file([
       '.babelrc',
       '.editorconfig',
       '.eslintrc.yml',
@@ -33,124 +63,63 @@ const files = {
       '.gitignore',
       '.yo-rc.json',
       'package.json'
-    ]
-  },
-  // Project application modules by base template
-  scripts: {
-    commonjs: [
-      'src/app.js',
+    ]);
+  }
+
+  function checkReadme() {
+    assert.fileContent([
+      ['README.md', `# [${title}]`],
+      ['README.md', `>   ${description}`]
+    ]);
+  }
+
+  function checkGulpTasks() {
+    assert.file([
+      'gulpfile.js/index.js',
+      'gulpfile.js/lib/bundler.js',
+      'gulpfile.js/lib/get-named-buffer.js',
+      'gulpfile.js/tasks/dev.js',
+      'gulpfile.js/tasks/dist.js'
+    ]);
+    assert.fileContent(
+      'gulpfile.js/config.js',
+      `{PHASER_BUILDS}/${customBuild}.js`
+    );
+  }
+
+  function checkCommonJsModules() {
+    assert.file([
       'src/assets.js',
       'src/objects/Logo.js',
       'src/states/Boot.js',
       'src/states/Game.js',
       'src/states/index.js',
       'src/states/Preloader.js'
-    ],
-    esnext: [
-      'src/app.js',
+    ]);
+    assert.fileContent('src/app.js', 'exports.init = function () {');
+  }
+
+  function checkECMAScriptModules() {
+    assert.file([
       'src/assets.js',
       'src/objects/Logo.js',
       'src/states.js',
       'src/states/Boot.js',
       'src/states/Game.js',
       'src/states/Preloader.js'
-    ]
-  },
-  // Common Gulp modules
-  gulpfiles: [
-    'gulpfile.js/config.js',
-    'gulpfile.js/index.js',
-    'gulpfile.js/lib/bundler.js',
-    'gulpfile.js/lib/get-named-buffer.js',
-    'gulpfile.js/tasks/dev.js',
-    'gulpfile.js/tasks/dist.js'
-  ]
-};
+    ]);
+    assert.fileContent('src/app.js', 'export function init() {');
+  }
 
-describe('default generator', function () {
-  this.timeout(0);
-
-  describe('CommonJS-based projects', () => {
-    before(() => helpers
-      .run(require.resolve('../generators/app'))
-      .withPrompts({
-        title: 'My Test Game',
-        description: 'My awesome test game.',
-        customBuild: 'phaser',
-        baseTemplate: 'commonjs'
-      })
-      .toPromise());
-
-    it('configuration files', () => assert.file(files.config.commonjs));
-
-    it('README', () => {
-      const file = 'README.md';
-      assert.file(file);
-      assert.fileContent(file, '# [My Test Game]');
-      assert.fileContent(file, '>   My awesome test game.');
-    });
-
-    it('gulp tasks', () => assert.file(files.gulpfiles));
-
-    it('Phaser build configured in tasks', () => {
-      assert.fileContent('gulpfile.js/config.js',
-        '{PHASER_BUILDS}/phaser.js');
-    });
-
-    it('sample game scripts', () => assert.file(files.scripts.commonjs));
-
-    it('sample game scripts are CommonJS modules', () => {
-      assert.fileContent('src/app.js', 'exports.init = function () {');
-    });
-
-    it('sample game assets', () => assert.file(files.assets));
-
-    it('sample `index.html` title and description', () => {
-      const file = 'static/index.html';
-      assert.fileContent(file, '<title>My Test Game</title>');
-      assert.fileContent(file, '<meta name="description" content="My awesome test game.">');
-    });
-  });
-
-  describe('ECMAScript-based projects', () => {
-    before(() => helpers
-      .run(require.resolve('../generators/app'))
-      .withPrompts({
-        title: 'My Test Game',
-        description: 'My awesome test game.',
-        customBuild: 'phaser',
-        baseTemplate: 'esnext'
-      })
-      .toPromise());
-
-    it('configuration files', () => assert.file(files.config.esnext));
-
-    it('README', () => {
-      const file = 'README.md';
-      assert.file(file);
-      assert.fileContent(file, '# [My Test Game]');
-      assert.fileContent(file, '>   My awesome test game.');
-    });
-
-    it('gulp tasks', () => assert.file(files.gulpfiles));
-
-    it('Phaser build configured in tasks', () => {
-      assert.fileContent('gulpfile.js/config.js',
-        '{PHASER_BUILDS}/phaser.js');
-    });
-
-    it('sample game scripts', () => assert.file(files.scripts.esnext));
-
-    it('sample game scripts are ECMAScript modules', () => {
-      assert.fileContent('src/app.js', 'export function init() {');
-    });
-
-    it('sample game assets', () => assert.file(files.assets));
-
-    it('sample `index.html` title and description', () => {
-      const file = 'static/index.html';
-      assert.fileContent(file, '<title>My Test Game</title>');
-      assert.fileContent(file, '<meta name="description" content="My awesome test game.">');
-    });
-  });
+  function checkAssets() {
+    assert.file([
+      'static/assets/phaser.png',
+      'static/assets/progress-bar.png',
+      'static/assets/splash-screen.png'
+    ]);
+    assert.fileContent([
+      ['static/index.html', `<title>${title}</title>`],
+      ['static/index.html', `<meta name="description" content="${description}">`]
+    ]);
+  }
 });